player position tracked with doubles, fish eye effect fixed

+++ cube3d.h
@@ -6,7 +6,7 @@
 /*   By: vguttenb <vguttenb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 12:54:37 by emgarcia          #+#    #+#             */
-/*   Updated: 2022/08/17 19:54:09 by vguttenb         ###   ########.fr       */
+/*   Updated: 2022/08/19 18:25:51 by vguttenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,8 +77,8 @@ typedef struct s_general
 	char	pos;
 	size_t	map_width;
 	size_t	map_height;
-	float	posx;
-	float	posy;
+	double	posx;
+	double	posy;
 	bool	game;
 	int		move_dir;
 	int		rotate_dir;
